✨  Introducing new features (SignUp Screen). 🚀Updating  files AppRegex to check the email and password validation. feat🎨 PasswordValidation. ♻️ Refactoring code.
