+++ lib/delete_this_after_merge/sign_up_widgets/email_and_password_validate.dart
@@ -1,87 +1,87 @@
-import 'package:flutter/material.dart';
-import 'package:shopping_app/delete_this_after_merge/sign_up_widgets/password_validation.dart';
-import 'package:shopping_app/delete_this_after_merge/widgets/app_regex.dart';
-import 'package:shopping_app/delete_this_after_merge/widgets/spacing.dart';
-import 'package:shopping_app/delete_this_after_merge/widgets/text_form_field.dart';
-class EmailAndPassword extends StatefulWidget {
-  const EmailAndPassword({super.key});
+// import 'package:flutter/material.dart';
+// import 'package:shopping_app/delete_this_after_merge/sign_up_widgets/password_validation.dart';
+// import 'package:shopping_app/core/helpers/app_regex.dart';
+// import 'package:shopping_app/delete_this_after_merge/widgets/spacing.dart';
+// import 'package:shopping_app/delete_this_after_merge/widgets/text_form_field.dart';
+// class EmailAndPassword extends StatefulWidget {
+//   const EmailAndPassword({super.key});
 
-  @override
-  State<EmailAndPassword> createState() => _EmailAndPasswordState();
-}
+//   @override
+//   State<EmailAndPassword> createState() => _EmailAndPasswordState();
+// }
 
-class _EmailAndPasswordState extends State<EmailAndPassword> {
-  //  we will use it in cubit later
-  final formKey = GlobalKey<FormState>();
-  //  we will use them in cubit later
-  TextEditingController emailController = TextEditingController();
-  TextEditingController passwordController = TextEditingController();
+// class _EmailAndPasswordState extends State<EmailAndPassword> {
+//   //  we will use it in cubit later
+//   final formKey = GlobalKey<FormState>();
+//   //  we will use them in cubit later
+//   TextEditingController emailController = TextEditingController();
+//   TextEditingController passwordController = TextEditingController();
 
-  //  to validate password
-  bool isObscureText = true;
-  bool hasLowerCase = false;
-  bool hasUpperCase = false;
-  bool hasSpecialCharacters = false;
-  bool hasNumber = false;
-  bool hasMinLength = false;
+//   //  to validate password
+//   bool isObscureText = true;
+//   bool hasLowerCase = false;
+//   bool hasUpperCase = false;
+//   bool hasSpecialCharacters = false;
+//   bool hasNumber = false;
+//   bool hasMinLength = false;
 
-  @override
-  void initState() {
-    super.initState();
-    passwordController;
-  }
+//   @override
+//   void initState() {
+//     super.initState();
+//     passwordController;
+//   }
 
-  @override
-  Widget build(BuildContext context) {
-    return Form(
-      key: formKey,
-      child: Column(
-        children: [
-          AppTextFormField(
-            // we will use controller from cubit
-            controller: emailController,
-            hintText: "E-mail",
-            validator: (value) {
-              if (value == null ||
-                  value.isEmpty ||
-                  !AppRegex.isEmailValid(value)) {
-                return 'Please enter a valid email';
-              }
-            },
-          ),
-          verticalSpace(18),
-          AppTextFormField(
-            // we will use controller from cubit
-            controller: passwordController,
-            hintText: 'Password',
-            isObscureText: isObscureText,
-            suffixIcon: GestureDetector(
-              onTap: () {
-                setState(() {
-                  isObscureText = !isObscureText;
-                });
-              },
-              child: Icon(
-                isObscureText ? Icons.visibility_off : Icons.visibility,
-              ),
-            ),
-            validator: (value) {
-              if (value == null || value.isEmpty) {
-                return 'Please enter a valid password';
-              }
-            },
-          ),
-          verticalSpace(24),
-          PasswordValidation(
-            hasLowerCase: hasLowerCase,
-            hasUpperCase: hasUpperCase,
-            hasSpecialCharacters: hasSpecialCharacters,
-            hasNumber: hasNumber,
-            hasMinLength: hasMinLength,
-            isObscureText: isObscureText,
-          ),
-        ],
-      ),
-    );
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     return Form(
+//       key: formKey,
+//       child: Column(
+//         children: [
+//           AppTextFormField(
+//             // we will use controller from cubit
+//             controller: emailController,
+//             hintText: "E-mail",
+//             validator: (value) {
+//               if (value == null ||
+//                   value.isEmpty ||
+//                   !AppRegex.isEmailValid(value)) {
+//                 return 'Please enter a valid email';
+//               }
+//             },
+//           ),
+//           verticalSpace(18),
+//           AppTextFormField(
+//             // we will use controller from cubit
+//             controller: passwordController,
+//             hintText: 'Password',
+//             isObscureText: isObscureText,
+//             suffixIcon: GestureDetector(
+//               onTap: () {
+//                 setState(() {
+//                   isObscureText = !isObscureText;
+//                 });
+//               },
+//               child: Icon(
+//                 isObscureText ? Icons.visibility_off : Icons.visibility,
+//               ),
+//             ),
+//             validator: (value) {
+//               if (value == null || value.isEmpty) {
+//                 return 'Please enter a valid password';
+//               }
+//             },
+//           ),
+//           verticalSpace(24),
+//           PasswordValidation(
+//             hasLowerCase: hasLowerCase,
+//             hasUpperCase: hasUpperCase,
+//             hasSpecialCharacters: hasSpecialCharacters,
+//             hasNumber: hasNumber,
+//             hasMinLength: hasMinLength,
+//             isObscureText: isObscureText,
+//           ),
+//         ],
+//       ),
+//     );
+//   }
+// }
