+++ lib/delete_this_after_merge/sign_up_widgets/password_validation.dart
@@ -5,7 +5,6 @@ import 'package:shopping_app/delete_this_after_merge/theming/text_styles.dart';
 import 'package:shopping_app/delete_this_after_merge/widgets/spacing.dart';
 
 class PasswordValidation extends StatelessWidget {
-  final bool isObscureText;
   final bool hasLowerCase;
   final bool hasUpperCase;
   final bool hasSpecialCharacters;
@@ -14,30 +13,11 @@ class PasswordValidation extends StatelessWidget {
 
   const PasswordValidation(
       {super.key,
-      required this.isObscureText,
       required this.hasLowerCase,
       required this.hasUpperCase,
       required this.hasSpecialCharacters,
       required this.hasNumber,
       required this.hasMinLength});
-
-  Widget buildValidationRow(String text, bool hasValidated) {
-    return Row(
-      children: [
-        const CircleAvatar(
-          radius: 2.5,
-          backgroundColor: ColorsManager.red,
-        ),
-        horizontalSpace(6),
-        Text(
-          text,
-          style: TextStyles.font12RedRegular.copyWith(
-            color: hasValidated ? ColorsManager.green : ColorsManager.red,
-          ),
-        )
-      ],
-    );
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -53,6 +33,26 @@ class PasswordValidation extends StatelessWidget {
         buildValidationRow('At least 1 number', hasNumber),
         verticalSpace(2),
         buildValidationRow('At least 8 characters long', hasMinLength),
+      ],
+    );
+  }
+
+  Widget buildValidationRow(String text, bool hasValidated) {
+    return Row(
+      children: [
+        const CircleAvatar(
+          radius: 2.5,
+          backgroundColor: ColorsManager.red,
+        ),
+        horizontalSpace(6),
+        Text(
+          text,
+          style: TextStyles.font12RedRegular.copyWith(
+            color: hasValidated
+                ? const Color(0xFF2F8655)
+                : const Color(0xFFDD2222),
+          ),
+        )
       ],
     );
   }
