+++ lib/features/sign_up/ui/sign_up_screen.dart
@@ -1,27 +1,12 @@
-// import 'package:flutter/material.dart';
-
-// class SignUpScreen extends StatelessWidget {
-//   const SignUpScreen({super.key});
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return const Scaffold(
-//       body: SafeArea(
-//         child: Center(
-//           child: Text('Sign Up Screen!!'),
-//         ),
-//       ),
-//     );
-//   }
-// }
-
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:shopping_app/delete_this_after_merge/sign_up_widgets/already_have_account.dart';
+import 'package:shopping_app/delete_this_after_merge/sign_up_widgets/password_validation.dart';
 import 'package:shopping_app/delete_this_after_merge/theming/text_styles.dart';
 import 'package:shopping_app/delete_this_after_merge/widgets/spacing.dart';
 import 'package:shopping_app/delete_this_after_merge/widgets/text_button.dart';
 import 'package:shopping_app/delete_this_after_merge/widgets/text_form_field.dart';
+import 'package:shopping_app/features/sign_up/ui/widgets/sign_up_form.dart';
 
 class SignUpScreen extends StatefulWidget {
   const SignUpScreen({super.key});
@@ -31,8 +16,7 @@ class SignUpScreen extends StatefulWidget {
 }
 
 class _SignUpScreenState extends State<SignUpScreen> {
-  final formKey = GlobalKey<FormState>();
-  bool isObscureText = true;
+  //final formKey = GlobalKey<FormState>();
 
   @override
   Widget build(BuildContext context) {
@@ -56,66 +40,7 @@ class _SignUpScreenState extends State<SignUpScreen> {
                   width: 240.w,
                 ),
                 verticalSpace(2),
-                Form(
-                  key: formKey,
-                  child: Column(
-                    children: [
-                      AppTextFormField(
-                        contentPadding: EdgeInsets.symmetric(
-                          horizontal: 20.w,
-                          vertical: 18.h,
-                        ),
-                        hintText: 'E-mail',
-                        inputTextStyle: TextStyles.font18BlackRegular,
-                        validator: (value) {},
-                      ),
-                      verticalSpace(16),
-                      AppTextFormField(
-                        contentPadding: EdgeInsets.symmetric(
-                          horizontal: 20.w,
-                          vertical: 18.h,
-                        ),
-                        hintText: 'Password',
-                        validator: (value) {},
-                        isObscureText: isObscureText,
-                        suffixIcon: GestureDetector(
-                          onTap: () {
-                            setState(() {
-                              isObscureText = !isObscureText;
-                            });
-                          },
-                          child: Icon(
-                            isObscureText
-                                ? Icons.visibility_off
-                                : Icons.visibility,
-                          ),
-                        ),
-                      ),
-                      verticalSpace(16),
-                      AppTextFormField(
-                        contentPadding: EdgeInsets.symmetric(
-                          horizontal: 20.w,
-                          vertical: 18.h,
-                        ),
-                        hintText: 'Confirm password',
-                        validator: (value) {},
-                        isObscureText: isObscureText,
-                        suffixIcon: GestureDetector(
-                          onTap: () {
-                            setState(() {
-                              isObscureText = !isObscureText;
-                            });
-                          },
-                          child: Icon(
-                            isObscureText
-                                ? Icons.visibility_off
-                                : Icons.visibility,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
+                const SignUpForm(),
                 verticalSpace(24),
                 AppTextButton(
                   buttonText: 'Sign Up',
